+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from './PlaywrightWorld';
-export * from './defineCucumber';
 export * from './filter';
 export * from './Template';
 export * from '@cucumber/cucumber';
+export * from './cucumberAdapter';