+++ test/cucumber.config.ts
@@ -1,4 +1,6 @@
-export default {
+import { defineConfig } from "../src";
+
+export default defineConfig({
     paths: ['test/features/*.feature'],
     require: ['test/step_definitions/*.ts']
-}
+})